agents/20-churn-intelligence-agent/requirements.txt
strands-agents>=0.1.0
boto3>=1.34.0
strands-agents-tools>=0.1.0
pandas>=2.0.0
scikit-learn>=1.3.0
